+++ src/components/VueBankCardBase.vue
@@ -55,11 +55,9 @@
                         ref="cardNumber"
                         placeholder="0000 0000 0000 0000"
                         v-mask="cardNumberMask"
-                        :value="cardNumber"
+                        v-model="cardNumber"
                         :class="fieldCssClasses('cardNumber')"
                         :readonly="!isNew"
-                        @input="$emit('input-card-number', $event.target.value)"
-                        @keydown.delete="moveCaretTo('back', 'cardNumber')"
                         @focus="clearErrors('cardNumber')"
                         @blur="$v.cardNumber.$touch()"
                     />
@@ -106,17 +104,8 @@
                                 ref="expDateMonth"
                                 placeholder="ММ"
                                 v-mask="expDateMonthMask"
-                                :value="expDateMonth"
+                                v-model="expDateMonth"
                                 :class="fieldCssClasses('expDateMonth')"
-                                @input="
-                                    $emit(
-                                        'input-exp-date-month',
-                                        $event.target.value
-                                    )
-                                "
-                                @keydown.delete="
-                                    moveCaretTo('back', 'expDateMonth')
-                                "
                                 @focus="clearErrors('expDateMonth')"
                                 @blur="
                                     autocompleteDate($event);
@@ -139,17 +128,8 @@
                                 ref="expDateYear"
                                 placeholder="ГГ"
                                 v-mask="expDateYearMask"
-                                :value="expDateYear"
+                                v-model="expDateYear"
                                 :class="fieldCssClasses('expDateYear')"
-                                @input="
-                                    $emit(
-                                        'input-exp-date-year',
-                                        $event.target.value
-                                    )
-                                "
-                                @keydown.delete="
-                                    moveCaretTo('back', 'expDateYear')
-                                "
                                 @focus="clearErrors('expDateYear')"
                                 @blur="
                                     autocompleteDate($event);
@@ -192,7 +172,7 @@
                         data-cp="cvv"
                         ref="cvv"
                         v-mask="cvvMask"
-                        :value="cvv"
+                        v-model="cvv"
                         :placeholder="cardInfo.codeName || 'CVV'"
                         :class="[
                             ...fieldCssClasses('cvv'),
@@ -201,8 +181,6 @@
                                     isCvvSecured && !isFieldEmpty('cvv')
                             }
                         ]"
-                        @input="$emit('input-cvv', $event.target.value)"
-                        @keydown.delete="moveCaretTo('back', 'cvv')"
                         @focus="
                             clearErrors('cvv');
                             isCvvSecured = false;
@@ -229,14 +207,9 @@
 <script>
 import { mask } from "vue-the-mask";
 import { validationMixin } from "vuelidate";
+
 import getBrands from "@/services/card-info/utils/get-brands";
-import {
-    commonMixin,
-    validatorsMixin,
-    moveCaretMixin,
-    helpersMixin
-} from "@/mixins";
-import clickOutside from "@/utils/click-outside-directive";
+import { commonMixin, validatorsMixin, helpersMixin } from "@/mixins";
 import VueBankCardTooltip from "./VueBankCardTooltip";
 
 export default {
@@ -244,41 +217,8 @@ export default {
     components: {
         VueBankCardTooltip
     },
-    directives: { mask, clickOutside },
-    mixins: [
-        commonMixin,
-        validationMixin,
-        validatorsMixin,
-        moveCaretMixin,
-        helpersMixin
-    ],
-    props: {
-        isNew: Boolean,
-        cardInfo: {
-            type: Object,
-            default: null
-        },
-        cardNumber: {
-            type: String,
-            required: true
-        },
-        cardHolderName: {
-            type: String,
-            required: true
-        },
-        expDateMonth: {
-            type: String,
-            required: true
-        },
-        expDateYear: {
-            type: String,
-            required: true
-        },
-        cvv: {
-            type: String,
-            required: true
-        }
-    },
+    directives: { mask },
+    mixins: [commonMixin, validationMixin, validatorsMixin, helpersMixin],
     data() {
         return {
             fields: [